chore: make SetPluginDir function more clear
