+++ internal/pkg/pluginengine/cmd_delete.go
@@ -7,7 +7,6 @@ import (
 	"github.com/devstream-io/devstream/internal/pkg/configmanager"
 	"github.com/devstream-io/devstream/internal/pkg/pluginmanager"
 	"github.com/devstream-io/devstream/internal/pkg/statemanager"
-	"github.com/devstream-io/devstream/pkg/util/file"
 	"github.com/devstream-io/devstream/pkg/util/interact"
 	"github.com/devstream-io/devstream/pkg/util/log"
 )
@@ -18,7 +17,7 @@ func Remove(configFile string, continueDirectly bool, isForceDelete bool) error 
 		return err
 	}
 
-	if err := file.SetPluginDir(cfg.PluginDir); err != nil {
+	if err = SetPluginDir(cfg.PluginDir); err != nil {
 		log.Errorf("Error: %s.", err)
 	}
