+++ internal/pkg/pluginengine/cmd_destroy.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/devstream-io/devstream/internal/pkg/configmanager"
 	"github.com/devstream-io/devstream/internal/pkg/statemanager"
-	"github.com/devstream-io/devstream/pkg/util/file"
 	"github.com/devstream-io/devstream/pkg/util/interact"
 	"github.com/devstream-io/devstream/pkg/util/log"
 )
@@ -21,7 +20,7 @@ func Destroy(configFile string, continueDirectly bool, isForceDestroy bool) erro
 		return fmt.Errorf("failed to load the config file")
 	}
 
-	if err := file.SetPluginDir(cfg.PluginDir); err != nil {
+	if err = SetPluginDir(cfg.PluginDir); err != nil {
 		log.Errorf("Error: %s.", err)
 	}
