--- internal/pkg/pluginengine/const.go
@@ -1,10 +1,0 @@
-package pluginengine
-
-type CommandType string
-
-const (
-	CommandApply  CommandType = "apply"
-	CommandDelete CommandType = "delete"
-
-	askUserIfContinue string = "Continue? [y/n]"
-)
