+++ internal/pkg/pluginengine/plugin.go
@@ -1,11 +1,16 @@
 package pluginengine
 
 import (
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/mitchellh/go-homedir"
 	"github.com/spf13/viper"
 
-	"github.com/devstream-io/devstream/internal/pkg/statemanager"
-
 	"github.com/devstream-io/devstream/internal/pkg/configmanager"
+	"github.com/devstream-io/devstream/internal/pkg/statemanager"
+	"github.com/devstream-io/devstream/pkg/util/log"
 )
 
 // DevStreamPlugin is a struct, on which Create/Read/Update/Delete interfaces are defined.
@@ -63,4 +68,51 @@ func Delete(tool *configmanager.Tool) (bool, error) {
 
 func renderInstanceIDtoOptions(tool *configmanager.Tool) {
 	tool.Options["instanceID"] = interface{}(tool.InstanceID)
+}
+
+func SetPluginDir(conf string) error {
+	pluginDir, err := getPluginDir(conf)
+	if err != nil {
+		return err
+	}
+	viper.Set("plugin-dir", pluginDir)
+	return nil
+}
+
+func getPluginDir(conf string) (string, error) {
+	pluginDir := viper.GetString("plugin-dir")
+	if pluginDir == "" {
+		pluginDir = conf
+	}
+
+	if pluginDir == "" {
+		homeDir, err := homedir.Dir()
+		if err != nil {
+			return "", err
+		}
+		defaultPluginDir := filepath.Join(homeDir, ".devstream", "plugins")
+		return defaultPluginDir, nil
+	}
+
+	pluginRealDir, err := handlePathWithHome(pluginDir)
+	if err != nil {
+		return "", err
+	}
+	return pluginRealDir, nil
+}
+
+// handlePathWithHome deal with "~" in the filePath
+func handlePathWithHome(filePath string) (string, error) {
+	if !strings.Contains(filePath, "~") {
+		return filePath, nil
+	}
+
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		return "", err
+	}
+	retPath := filepath.Join(homeDir, strings.TrimPrefix(filePath, "~"))
+	log.Debugf("real path: %s.", retPath)
+
+	return retPath, nil
 }
